refactor: type-only import 추가 및 eslint 충돌 해결
