+++ apis/auth/mutations.ts
@@ -1,6 +1,4 @@
 import {
-  ReissueRequest,
-  ReissueResponse,
   postEmail,
   postEmailVerify,
   postLogin,
