+++ app/grouping/components/GroupingCard.client.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import React from 'react';
 
-import type { Grouping } from '@/apis/groups';
+import type { Grouping } from '@/apis/groups/type';
 
 interface GroupingCardProps {
   groupingData: Grouping;
